refactor svg forms to come from polygons cmp
