+++ ui/footer.module.scss
@@ -1,17 +1,8 @@
 .divider_override {
   margin: 2rem 0;
 
-  .footer_separator_polygons {
-    width: 40vw;
-    padding: 0;
-  }
-
   @media only screen and (min-width: 768px) {
     margin: 0;
-
-    .footer_separator_polygons {
-      width: 25vw;
-    }
   }
 }
 
@@ -84,14 +75,6 @@
           font-size: 1rem;
         }
       }
-    }
-  }
-
-  .footer_polygons {
-    height: 200px;
-
-    @media only screen and (min-width: 768px) {
-      display: none;
     }
   }
 }
