+++ ui/polygons.tsx
@@ -1,0 +1,256 @@
+import { useEffect, useState } from "react";
+import Link from 'next/link';
+
+import styles from './polygons.module.scss';
+
+export function getTrianglesWithColorAndFill(paths: string[]) {
+  const [isMounted, setIsMounted] = useState(false);
+
+  useEffect(() => {
+    setIsMounted(true);
+  }, []);
+
+  return paths.map((path, index) => {
+    const color = Math.random() < 0.5 ? "var(--primary, #07BDBD)" : "var(--secondary, #9E19EF)";
+    const hasFill = Math.random() < 0.5;
+    return (
+      <>
+        {isMounted && <path key={index} d={path} stroke={color} fill={hasFill ? color : "var(--background)"} />}
+      </>
+    );
+  });
+}
+
+function HeaderBackgroundDesktop() {
+  const trianglePaths = [
+    "M1284.5 206.5H1351.79L1284.5 273.793V206.5Z",
+    "M1387.21 250.702L1434.08 297.579H1340.33L1387.21 250.702Z",
+    "M1434.5 425.5L1389.21 425.5L1434.5 380.207L1434.5 425.5Z",
+    "M1481.5 425.5L1481.5 380.207L1526.79 425.5L1481.5 425.5Z",
+    "M1587.5 363.554L1522.95 299L1652.05 299L1587.5 363.554Z",
+    "M1481.5 471.5L1481.5 449.207L1503.79 471.5L1481.5 471.5Z",
+    "M1437.5 297.5L1437.5 275.207L1459.79 297.5L1437.5 297.5Z",
+    "M1432.5 208.5L1432.5 230.793L1410.21 208.5L1432.5 208.5Z",
+    "M1435.5 205.5L1435.5 161.207L1479.79 205.5L1435.5 205.5Z",
+    "M1679.5 45.5L1635.21 45.5L1679.5 1.20711L1679.5 45.5Z",
+    "M1679.5 57.5V92.7929L1644.21 57.5L1679.5 57.5Z",
+    "M1540.5 80.5L1473.21 80.5L1540.5 13.2071L1540.5 80.5Z",
+    "M1587.5 126.5L1587.5 81.2071L1632.79 126.5L1587.5 126.5Z",
+    "M1541.5 127.5H1563.79L1541.5 149.793V127.5Z",
+    "M355.753 321.122L318.069 358.806V283.438L355.753 321.122Z",
+    "M345.5 225.5L238.207 225.5L345.5 118.207L345.5 225.5Z",
+    "M131.552 63.7071L162.884 95.0397L99.538 95.7209L131.552 63.7071Z",
+    "M257.291 108.998L238.722 127.567L238.319 90.0254L257.291 108.998Z",
+    "M65.791 128.761L97.8179 160.788L33.7641 160.788L65.791 128.761Z",
+    "M86.6172 281.527L34.7913 229.701L138.443 229.701L86.6172 281.527Z",
+    "M33.2637 96.234L49.0271 111.997L17.5002 111.997L33.2637 96.234Z",
+    "M188.12 186.037L203.884 201.8L172.357 201.8L188.12 186.037Z",
+    "M351.027 225.264L351.027 180.971L395.32 225.264L351.027 225.264Z",
+    "M0.707107 45.5269L32.0269 14.2071L32.0269 76.8467L0.707107 45.5269Z",
+    "M185.999 528.983L161.043 504.028H210.955L185.999 528.983Z",
+    "M188.024 275.432L235.607 227.849L235.607 323.015L188.024 275.432Z",
+    "M202.263 415.139L233.583 446.459L170.943 446.459L202.263 415.139Z",
+    "M234.082 381.905L219.733 396.254V367.556L234.082 381.905Z"
+  ]
+
+  return (
+    <Link href={typeof window !== 'undefined' ? window.location.href : "/"}>
+      <svg viewBox="0 0 1680 570" fill="none" xmlns="http://www.w3.org/2000/svg">
+        {getTrianglesWithColorAndFill(trianglePaths)}
+
+        <Link href="/">
+          <svg>
+            <rect x="700" y="175" width="270" height="390" opacity="0" fill="white" />
+            <path d="M963.975 464.998H744.794V242.76H963.975V464.998ZM745.799 463.981H962.97V463.979H962.972V243.779H962.97V243.777H745.799V243.779H745.797V463.979H745.799V463.981Z" fill="black" stroke="black" strokeWidth="0.00352113" />
+            <path d="M912.809 342.404C904.798 343.037 898.075 345.916 893.072 350.868C887.915 355.963 885.459 362.192 885.459 370.085C885.459 381.888 890.319 390.139 901.145 396.684C903.897 398.338 907.652 400.135 914.14 402.923C922.98 406.711 926.265 408.508 928.362 410.754C930.951 413.521 931.667 417.953 930.04 421.333C929.723 421.986 928.894 423.089 928.158 423.834C925.651 426.356 921.711 427.684 916.81 427.684C908.543 427.684 902.373 424.294 897.41 416.972C896.735 416.002 896.121 415.196 896.029 415.196C895.834 415.196 879.698 424.488 879.36 424.804C879.146 424.998 879.32 425.437 880.148 426.795C886.595 437.629 896.663 443.98 910.282 445.797C913.843 446.287 920.524 446.267 923.932 445.797C932.496 444.593 938.922 441.754 943.905 436.934C949.205 431.839 951.814 425.233 951.814 416.962C951.814 411.673 950.945 407.548 948.97 403.464C947.609 400.666 946.258 398.787 943.925 396.418C939.454 391.895 933.663 388.464 922.11 383.512C913.372 379.765 910.446 378.161 908.522 376.109C906.68 374.138 905.892 371.831 906.066 368.962C906.22 366.674 906.854 365.133 908.246 363.632C910.262 361.457 912.595 360.538 916.196 360.568C920.289 360.568 922.939 361.63 925.63 364.336C926.582 365.265 927.81 366.766 928.383 367.675C929.038 368.717 929.529 369.278 929.692 369.217C930.245 368.982 945.655 359.037 945.655 358.894C945.655 358.812 944.867 357.587 943.915 356.147C941.879 353.104 938.216 349.296 935.689 347.621C931.023 344.527 926.04 342.945 919.471 342.455C916.268 342.169 915.848 342.169 912.809 342.404Z" fill="black" />
+            <path d="M846.812 380.847L846.751 418.259L846.28 420.015C845.666 422.323 844.94 423.701 843.671 424.967C841.891 426.744 839.599 427.49 835.782 427.51C830.328 427.53 827.064 425.223 823.268 418.749C822.654 417.738 822.102 416.911 822.04 416.932C821.958 416.973 818.121 419.28 813.517 422.047L805.137 427.122L805.771 428.388C809.567 435.944 816.136 441.652 824.046 444.276C832.395 447.043 843.129 446.798 850.997 443.643C860.431 439.834 865.967 431.962 867.43 420.26C867.553 419.076 867.645 405.853 867.645 380.969V343.455H846.894L846.812 380.847Z" fill="black" />
+            <path d="M745.779 289.094L745.779 243.744L768.502 266.419L745.779 289.094Z" fill="black" />
+            <path d="M768.502 266.419L768.502 311.769L745.779 289.094L768.502 266.419Z" fill="black" />
+            <path d="M836.669 289.094L791.224 289.094L813.946 266.419L836.669 289.094Z" fill="black" />
+            <path d="M859.392 311.769L813.947 311.769L836.669 289.094L859.392 311.769Z" fill="black" />
+            <path d="M882.114 243.744V289.094L859.392 266.419L882.114 243.744Z" fill="black" />
+            <path d="M836.669 243.744V289.094L813.946 266.419L836.669 243.744Z" fill="black" />
+            <path d="M791.224 289.094V334.444L768.502 311.769L791.224 289.094Z" fill="black" />
+            <path d="M813.947 311.769V357.118L791.225 334.443L813.947 311.769Z" fill="black" />
+            <path d="M745.779 334.443L791.224 334.443L768.502 357.118L745.779 334.443Z" fill="black" />
+            <path d="M745.779 379.793L791.224 379.793L768.502 402.468L745.779 379.793Z" fill="black" />
+            <path d="M744.792 289.516L698.5 289.516L721.646 266.419L744.792 289.516Z" fill="black" />
+            <path d="M698.5 196L744.792 196L721.646 219.238L698.5 196Z" fill="black" />
+            <path d="M768.22 219.238L814.229 219.238L791.224 242.899L768.22 219.238Z" fill="black" />
+            <path d="M836.67 196.563L836.67 242.758L813.947 219.661L836.67 196.563Z" fill="black" />
+            <path d="M721.646 219.238L721.646 266.419L698.5 242.829L721.646 219.238Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M744.359 242.179L744.933 242.758L768.22 219.238L721.646 219.238L722.498 220.099L766.283 220.099L744.359 242.179Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M880.875 242.758H882.114L859.391 219.661L836.669 242.758H838.184L839.176 241.749H839.15L859.391 221.073L859.454 221.136L859.529 221.061L880.875 242.758Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M744.791 290.331L744.791 289.094L721.646 311.769L744.791 334.443L744.791 332.932L743.781 331.942L743.781 331.968L723.061 311.769L723.124 311.706L723.048 311.631L744.791 290.331Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M744.791 335.68L744.791 334.444L721.646 357.118L744.791 379.793L744.791 378.282L743.781 377.292L743.781 377.318L723.061 357.118L723.124 357.056L723.048 356.981L744.791 335.68Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M721.645 265.062L721.645 266.348L744.791 242.758L721.645 219.168L721.645 220.741L722.656 221.771L722.656 221.743L743.376 242.758L743.312 242.824L743.388 242.901L721.645 265.062Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M812.425 219.238H813.665L790.942 196.141L768.22 219.238H769.735L770.727 218.23H770.7L790.942 197.553L791.005 197.617L791.08 197.541L812.425 219.238Z" fill="black" />
+            <path d="M768.219 172.903L768.219 219.238L744.791 196.07L768.219 172.903ZM767.196 216.709L767.196 175.23L746.223 196.07L767.196 216.709Z" fill="black" />
+            <path d="M813.947 266.419H768.502L791.224 243.744L813.947 266.419ZM770.983 265.429H811.665L791.224 245.13L770.983 265.429Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M883.353 243.744L882.114 243.744L904.836 266.419L927.559 243.744L926.044 243.744L925.051 244.734L925.078 244.734L904.836 265.033L904.773 264.97L904.698 265.045L883.353 243.744Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M836.669 287.857L836.669 289.094L859.392 266.419L836.669 243.744L836.669 245.256L837.661 246.246L837.661 246.219L858.002 266.419L857.94 266.482L858.014 266.556L836.669 287.857Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M747.186 334.444L769.173 312.439L768.502 311.769L745.779 334.444H747.186Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M768.502 267.964L769.494 268.951L769.494 268.894L789.835 289.094L789.787 289.142L790.483 289.834L791.225 289.094L768.502 266.419L768.502 267.964Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M791.224 290.639L792.216 291.626L792.216 291.569L812.557 311.769L812.508 311.817L813.204 312.509L813.946 311.769L791.224 289.094L791.224 290.639Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M769.177 356.445L768.502 357.118L791.224 379.793L791.224 334.443L790.236 335.429L790.236 377.399L769.177 356.445Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M767.867 356.485L745.78 378.526L745.78 379.793L768.502 357.118L767.867 356.485Z" fill="black" />
+            <path d="M745.322 519V487.92H755.57C761.954 487.92 766.826 489.81 766.826 495.606C766.826 498.462 765.272 501.36 762.626 502.326V502.494C765.944 503.292 768.422 505.686 768.422 509.844C768.422 516.102 763.214 519 756.41 519H745.322ZM751.538 500.562H755.15C759.098 500.562 760.778 499.008 760.778 496.488C760.778 493.758 758.972 492.708 755.234 492.708H751.538V500.562ZM751.538 514.212H755.864C760.064 514.212 762.374 512.7 762.374 509.508C762.374 506.484 760.148 505.182 755.864 505.182H751.538V514.212ZM786.258 519.588C778.656 519.588 774.162 515.346 774.162 505.056V487.92H780.336V505.602C780.336 512.028 782.604 514.212 786.258 514.212C789.912 514.212 792.306 512.028 792.306 505.602V487.92H798.27V505.056C798.27 515.346 793.86 519.588 786.258 519.588ZM806.168 519V487.92H814.694C824.186 487.92 830.024 492.834 830.024 503.334C830.024 513.792 824.186 519 815.03 519H806.168ZM812.384 514.002H814.274C820.112 514.002 823.682 510.894 823.682 503.334C823.682 495.774 820.112 492.876 814.274 492.876H812.384V514.002ZM831.883 519L841.837 487.92H849.145L859.099 519H852.589L848.137 502.83C847.213 499.596 846.331 495.984 845.449 492.666H845.281C844.441 496.026 843.559 499.596 842.635 502.83L838.183 519H831.883ZM838.057 511.02V506.232H852.841V511.02H838.057ZM863.199 519V487.92H873.405C880.251 487.92 885.543 490.314 885.543 497.706C885.543 504.804 880.251 507.954 873.573 507.954H869.415V519H863.199ZM869.415 503.04H873.153C877.437 503.04 879.495 501.234 879.495 497.706C879.495 494.052 877.227 492.876 872.943 492.876H869.415V503.04ZM891.633 519V487.92H910.701V493.086H897.849V500.268H908.727V505.518H897.849V513.792H911.163V519H891.633ZM927.175 519.588C922.975 519.588 918.859 517.992 915.835 515.094L919.405 510.81C921.589 512.868 924.571 514.212 927.301 514.212C930.535 514.212 932.215 512.826 932.215 510.684C932.215 508.374 930.325 507.618 927.553 506.442L923.395 504.678C920.287 503.418 917.095 500.856 917.095 496.362C917.095 491.28 921.589 487.374 927.847 487.374C931.459 487.374 935.071 488.802 937.591 491.364L934.483 495.27C932.467 493.632 930.409 492.708 927.847 492.708C925.117 492.708 923.395 493.926 923.395 495.942C923.395 498.168 925.621 499.008 928.267 500.058L932.341 501.738C936.079 503.292 938.557 505.728 938.557 510.18C938.557 515.262 934.315 519.588 927.175 519.588ZM949.506 519V493.086H940.728V487.92H964.5V493.086H955.722V519H949.506Z" fill="black" />
+            <path d="M748.304 562.378C745.388 562.378 743.444 561.217 742.094 558.895L744.848 556.843C745.604 558.301 746.657 558.922 747.791 558.922C749.573 558.922 750.491 558.004 750.491 555.25V542.02H754.487V555.574C754.487 559.327 752.57 562.378 748.304 562.378ZM763.488 562.378C760.221 562.378 758.79 560.137 758.79 556.33V546.907H762.759V555.844C762.759 558.166 763.407 558.976 764.892 558.976C766.134 558.976 766.944 558.436 767.97 557.086V546.907H771.912V562H768.672L768.375 559.84H768.294C766.971 561.379 765.54 562.378 763.488 562.378ZM776.17 562V546.907H779.383L779.707 548.905H779.788C781.084 547.609 782.65 546.529 784.702 546.529C787.942 546.529 789.373 548.797 789.373 552.577V562H785.404V553.09C785.404 550.714 784.729 549.931 783.271 549.931C782.029 549.931 781.246 550.525 780.139 551.605V562H776.17ZM800.055 562.378C795.816 562.378 792.522 559.462 792.522 554.467C792.522 549.553 795.978 546.529 799.542 546.529C803.727 546.529 805.86 549.553 805.86 553.738C805.86 554.494 805.779 555.196 805.698 555.574H796.356C796.707 558.058 798.327 559.327 800.568 559.327C801.783 559.327 802.863 558.949 803.97 558.274L805.293 560.731C803.808 561.757 801.918 562.378 800.055 562.378ZM796.302 552.928H802.485C802.485 550.876 801.567 549.58 799.623 549.58C798.03 549.58 796.626 550.714 796.302 552.928ZM814.274 562V559.732C819.485 554.737 822.914 551.362 822.914 548.446C822.914 546.475 821.861 545.287 819.944 545.287C818.459 545.287 817.298 546.286 816.272 547.366L814.085 545.206C815.921 543.235 817.73 542.182 820.457 542.182C824.21 542.182 826.748 544.558 826.748 548.257C826.748 551.632 823.589 555.169 820.025 558.895C821.024 558.787 822.347 558.652 823.292 558.652H827.666V562H814.274ZM837.57 562.378C833.763 562.378 830.361 559.489 830.361 552.793C830.361 545.368 834.087 542.182 838.11 542.182C840.567 542.182 842.349 543.208 843.537 544.423L841.431 546.826C840.729 546.016 839.514 545.395 838.353 545.395C835.95 545.395 834.006 547.231 834.006 552.793C834.006 557.626 835.518 559.381 837.462 559.381C839.001 559.381 840.216 558.22 840.216 555.925C840.216 553.738 839.082 552.739 837.354 552.739C836.274 552.739 835.032 553.333 834.006 555.034L833.844 552.334C834.924 550.822 836.733 549.931 838.083 549.931C841.431 549.931 843.834 551.848 843.834 555.925C843.834 559.894 840.972 562.378 837.57 562.378ZM846.287 555.709V552.874H853.685V555.709H846.287ZM856.118 562V559.732C861.329 554.737 864.758 551.362 864.758 548.446C864.758 546.475 863.705 545.287 861.788 545.287C860.303 545.287 859.142 546.286 858.116 547.366L855.929 545.206C857.765 543.235 859.574 542.182 862.301 542.182C866.054 542.182 868.592 544.558 868.592 548.257C868.592 551.632 865.433 555.169 861.869 558.895C862.868 558.787 864.191 558.652 865.136 558.652H869.51V562H856.118ZM878.848 562.378C875.014 562.378 872.179 560.191 872.179 557.059C872.179 554.629 873.772 553.09 875.581 552.118V552.01C874.096 550.903 872.935 549.472 872.935 547.447C872.935 544.207 875.5 542.182 878.983 542.182C882.52 542.182 884.788 544.288 884.788 547.474C884.788 549.364 883.492 550.93 882.223 551.821V551.956C884.032 552.982 885.571 554.44 885.571 557.059C885.571 560.083 882.871 562.378 878.848 562.378ZM880.117 550.876C881.035 549.877 881.548 548.851 881.548 547.717C881.548 546.097 880.549 544.963 878.902 544.963C877.525 544.963 876.445 545.827 876.445 547.447C876.445 549.229 878.038 550.039 880.117 550.876ZM878.929 559.57C880.657 559.57 881.872 558.625 881.872 556.897C881.872 554.953 879.982 554.278 877.444 553.198C876.364 554.035 875.581 555.223 875.581 556.6C875.581 558.409 877.066 559.57 878.929 559.57ZM889.077 567.778L888.267 565.618C890.373 564.835 891.534 563.404 891.507 561.811L891.372 559.003L892.776 561.325C892.371 561.73 891.831 561.919 891.237 561.919C889.941 561.919 888.753 561.055 888.753 559.57C888.753 558.139 889.914 557.221 891.345 557.221C893.154 557.221 894.18 558.679 894.18 561.109C894.18 564.295 892.371 566.698 889.077 567.778ZM902.867 562V559.732C908.078 554.737 911.507 551.362 911.507 548.446C911.507 546.475 910.454 545.287 908.537 545.287C907.052 545.287 905.891 546.286 904.865 547.366L902.678 545.206C904.514 543.235 906.323 542.182 909.05 542.182C912.803 542.182 915.341 544.558 915.341 548.257C915.341 551.632 912.182 555.169 908.618 558.895C909.617 558.787 910.94 558.652 911.885 558.652H916.259V562H902.867ZM925.624 562.378C921.493 562.378 918.847 558.814 918.847 552.199C918.847 545.584 921.493 542.182 925.624 542.182C929.701 542.182 932.374 545.611 932.374 552.199C932.374 558.814 929.701 562.378 925.624 562.378ZM925.624 559.273C927.352 559.273 928.621 557.626 928.621 552.199C928.621 546.772 927.352 545.26 925.624 545.26C923.896 545.26 922.6 546.772 922.6 552.199C922.6 557.626 923.896 559.273 925.624 559.273ZM934.719 562V559.732C939.93 554.737 943.359 551.362 943.359 548.446C943.359 546.475 942.306 545.287 940.389 545.287C938.904 545.287 937.743 546.286 936.717 547.366L934.53 545.206C936.366 543.235 938.175 542.182 940.902 542.182C944.655 542.182 947.193 544.558 947.193 548.257C947.193 551.632 944.034 555.169 940.47 558.895C941.469 558.787 942.792 558.652 943.737 558.652H948.111V562H934.719ZM958.583 562V550.255C958.583 549.094 958.691 547.231 958.745 546.07H958.637C958.124 547.123 957.584 548.122 957.017 549.229L953.939 554.116H964.658V557.167H950.078V554.44L957.503 542.533H962.282V562H958.583Z" fill="black" />
+          </svg>
+        </Link>
+      </svg>
+    </Link>
+
+  )
+}
+
+function HeaderBackgroundMobile() {
+  const trianglePaths = [
+    "M290.595 41.3831L290.771 41.5599L290.948 41.3831L302.166 30.1657L302.592 29.739H301.989H279.554H278.951L279.377 30.1657L290.595 41.3831Z",
+    "M11.8572 3.78115L11.6804 3.60437L11.5036 3.78115L3.19519 12.0897L2.76184 12.523L3.37466 12.5164L19.8147 12.3397L20.4092 12.3333L19.9888 11.9129L11.8572 3.78115Z",
+    "M6.91575 25.895L7.09253 25.7182L6.91575 25.5414L3.72864 22.3543L3.29529 21.9209L3.30188 22.5337L3.36969 28.8402L3.37608 29.4347L3.79645 29.0143L6.91575 25.895Z",
+    "M11.4549 47.1091L11.2781 46.9323L11.1013 47.1091L2.21289 55.9976L1.77954 56.431L2.39235 56.4244L19.9801 56.2353L20.5745 56.2289L20.1542 55.8085L11.4549 47.1091Z",
+    "M30.6179 6.91003V6.66003H30.3679H23.2288H22.616L23.0539 7.0887L30.1931 14.076L30.6179 14.4918V13.8973V6.91003Z",
+    "M321.876 79.8207L322.053 79.6439L321.876 79.4671L305.259 62.8501L304.832 62.4234V63.0269V96.2609V96.8645L305.259 96.4377L321.876 79.8207Z",
+    "M318.013 69.9877L318.189 70.1645L318.366 69.9877L325.869 62.4853L326.295 62.0585H325.692H310.687H310.083L310.51 62.4853L318.013 69.9877Z",
+    "M358.995 83.0719V82.8219H358.745H347.245H346.659L347.065 83.2449L358.565 95.2449L358.995 95.6941V95.0719V83.0719Z",
+    "M261.585 52.6197L261.408 52.7965L261.585 52.9733L270.07 61.4586L270.51 61.8985L270.497 61.2765L270.143 44.6595L270.131 44.0737L269.717 44.488L261.585 52.6197Z",
+    "M305.7 28.2634V28.5134H305.95H328.95H329.554L329.127 28.0867L306.127 5.08663L305.7 4.65985V5.26341V28.2634Z",
+    "M399.01 125.022L399.187 125.199L399.363 125.022L412.445 111.941L412.872 111.514H412.268H386.105H385.501L385.928 111.941L399.01 125.022Z",
+    "M363.74 96.9661V97.2161H363.99H398.49H399.094L398.667 96.7893L364.167 62.2893L363.74 61.8625V62.4661V96.9661Z",
+    "M113.633 13.3049L113.809 13.1281L113.986 13.3049L123.074 22.3926L123.501 22.8193H122.897H104.722H104.118L104.545 22.3926L113.633 13.3049Z",
+    "M124.633 15.3337L124.456 15.1569L124.633 14.9801L131.132 8.48019L131.559 8.05341V8.65696V21.6569V22.2604L131.132 21.8336L124.633 15.3337Z",
+    "M139.986 20.7301L140.163 20.9069L139.986 21.0837L136.236 24.8336L135.809 25.2604V24.6569V17.157V16.5534L136.236 16.9802L139.986 20.7301Z",
+    "M145.611 36.7087L145.434 36.8855L145.258 36.7087L137.383 28.8337L136.956 28.4069H137.559H153.309H153.913L153.486 28.8337L145.611 36.7087Z",
+    "M155.218 34.6237L155.041 34.447L155.218 34.2702L158.271 31.2172L158.697 30.7904V31.3939V37.5V38.1035L158.271 37.6768L155.218 34.6237Z",
+  ]
+
+  return (
+    <Link href={typeof window !== 'undefined' ? window.location.href : "/"}>
+      <svg viewBox="0 0 415 170" fill="none" xmlns="http://www.w3.org/2000/svg">
+        {getTrianglesWithColorAndFill(trianglePaths)}
+
+        {/* menu background */}
+        <path d="M414.5 1.00001V0.500005H414L309 0.5H307.793L308.646 1.35355L413.646 106.354L414.5 107.207V106V1.00001Z" fill="#07BDBD" stroke="#07BDBD" />
+        {/* <path d="M368 19C368 17.3431 369.269 16 370.833 16H399.167C400.731 16 402 17.3431 402 19C402 20.6569 400.731 22 399.167 22H370.833C369.269 22 368 20.6569 368 19Z" fill="white" />
+            <path d="M368 30C368 28.3431 369.269 27 370.833 27H399.167C400.731 27 402 28.3431 402 30C402 31.6569 400.731 33 399.167 33H370.833C369.269 33 368 31.6569 368 30Z" fill="white" />
+            <path d="M368 41C368 39.3431 369.269 38 370.833 38H399.167C400.731 38 402 39.3431 402 41C402 42.6569 400.731 44 399.167 44H370.833C369.269 44 368 42.6569 368 41Z" fill="white" /> */}
+
+        <Link href="/">
+          <svg>
+            <rect x="28" y="24" width="96" height="141" opacity="0" fill="white" />
+            <path d="M122.365 129.812H40.172V46.4725H122.365V129.812ZM40.5489 129.43H121.988H121.989V46.8546H121.988L40.5489 46.854V129.43Z" fill="black" stroke="black" strokeWidth="0.00132042" />
+            <path d="M103.179 83.8388C100.175 84.0762 97.654 85.1559 95.7777 87.013C93.8438 88.9237 92.9229 91.2594 92.9229 94.2193C92.9229 98.6456 94.7455 101.739 98.8051 104.194C99.8373 104.814 101.246 105.488 103.678 106.533C106.993 107.954 108.225 108.628 109.012 109.47C109.982 110.508 110.251 112.17 109.641 113.437C109.522 113.682 109.211 114.096 108.935 114.375C107.995 115.321 106.518 115.819 104.68 115.819C101.579 115.819 99.2656 114.548 97.4046 111.802C97.1514 111.438 96.9211 111.136 96.8866 111.136C96.8137 111.136 90.7627 114.62 90.636 114.739C90.5555 114.812 90.6207 114.976 90.9315 115.486C93.3488 119.548 97.1245 121.93 102.232 122.611C103.567 122.795 106.073 122.788 107.35 122.611C110.562 122.16 112.972 121.095 114.84 119.288C116.828 117.377 117.806 114.9 117.806 111.798C117.806 109.815 117.48 108.268 116.74 106.736C116.229 105.687 115.723 104.983 114.848 104.094C113.171 102.398 110.999 101.111 106.667 99.2544C103.39 97.8492 102.293 97.248 101.572 96.4784C100.881 95.7394 100.586 94.874 100.651 93.7981C100.708 92.9404 100.946 92.3622 101.468 91.7993C102.224 90.9837 103.099 90.6391 104.449 90.6506C105.984 90.6506 106.978 91.0488 107.987 92.0635C108.344 92.412 108.805 92.9748 109.019 93.3156C109.265 93.7062 109.449 93.9168 109.511 93.8938C109.718 93.8057 115.496 90.0763 115.496 90.0227C115.496 89.992 115.201 89.5325 114.844 88.9926C114.08 87.8516 112.707 86.4234 111.759 85.7954C110.009 84.6352 108.141 84.0417 105.677 83.8579C104.476 83.7507 104.319 83.7507 103.179 83.8388Z" fill="black" />
+            <path d="M78.4301 98.2551L78.4071 112.285L78.2306 112.943C78.0004 113.809 77.7279 114.325 77.2521 114.8C76.5845 115.467 75.725 115.746 74.2938 115.754C72.2486 115.761 71.0246 114.896 69.601 112.468C69.3708 112.089 69.1636 111.779 69.1406 111.787C69.1099 111.802 67.671 112.668 65.9443 113.705L62.8018 115.608L63.0397 116.083C64.4632 118.916 66.9266 121.057 69.8926 122.041C73.0237 123.079 77.0488 122.987 79.9995 121.804C83.5372 120.375 85.6131 117.423 86.1618 113.035C86.2078 112.591 86.2424 107.632 86.2424 98.301V84.2332H78.4608L78.4301 98.2551Z" fill="black" />
+            <path d="M40.5426 63.8477V46.8416L49.0635 55.3446L40.5426 63.8477Z" fill="black" />
+            <path d="M49.0634 55.3447V72.3508L40.5425 63.8477L49.0634 55.3447Z" fill="black" />
+            <path d="M74.6265 63.8476H57.5846L66.1055 55.3445L74.6265 63.8476Z" fill="black" />
+            <path d="M83.1467 72.3507H66.1049L74.6258 63.8476L83.1467 72.3507Z" fill="black" />
+            <path d="M91.6677 46.8414V63.8476L83.1467 55.3445L91.6677 46.8414Z" fill="black" />
+            <path d="M74.6257 46.8414V63.8476L66.1047 55.3445L74.6257 46.8414Z" fill="black" />
+            <path d="M57.5844 63.8477V80.8539L49.0635 72.3508L57.5844 63.8477Z" fill="black" />
+            <path d="M66.1054 72.3506V89.3568L57.5845 80.8537L66.1054 72.3506Z" fill="black" />
+            <path d="M40.5425 80.8537H57.5843L49.0634 89.3568L40.5425 80.8537Z" fill="black" />
+            <path d="M40.5425 97.8599H57.5843L49.0634 106.363L40.5425 97.8599Z" fill="black" />
+            <path d="M40.1724 64.0061H22.813L31.4927 55.3445L40.1724 64.0061Z" fill="black" />
+            <path d="M22.8125 28.9375H40.1719L31.4922 37.6518L22.8125 28.9375Z" fill="black" />
+            <path d="M48.9578 37.6518H66.2113L57.5845 46.5246L48.9578 37.6518Z" fill="black" />
+            <path d="M74.6261 29.1487V46.4718L66.1052 37.8103L74.6261 29.1487Z" fill="black" />
+            <path d="M31.4922 37.6518V55.3445L22.8125 46.4982L31.4922 37.6518Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M40.0095 46.2547L40.2246 46.4719L48.9572 37.6519H31.4919L31.8114 37.9746H48.2311L40.0095 46.2547Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M91.2037 46.4718H91.6683L83.1474 37.8103L74.6265 46.4718H75.1947L75.5667 46.0936H75.5567L83.1474 38.3398L83.1711 38.3638L83.1992 38.3352L91.2037 46.4718Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M40.1724 64.3114V63.8477L31.4927 72.3508L40.1724 80.8539V80.2872L39.7933 79.9159V79.9256L32.0233 72.3508L32.0472 72.3273L32.0186 72.2993L40.1724 64.3114Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M40.1719 81.3175V80.8538L31.4922 89.3569L40.1719 97.86V97.2933L39.7929 96.922V96.9317L32.0228 89.3569L32.0467 89.3334L32.0181 89.3054L40.1719 81.3175Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M31.4912 54.8359V55.3182L40.1709 46.4719L31.4912 37.6255V38.2151L31.8702 38.6014V38.5912L39.6403 46.4719L39.6164 46.4963L39.645 46.5255L31.4912 54.8359Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M65.5346 37.6519H65.9991L57.4782 28.9904L48.9573 37.6519H49.5255L49.8976 37.2737H49.8875L57.4782 29.5199L57.5019 29.5439L57.5301 29.5153L65.5346 37.6519Z" fill="black" />
+            <path d="M48.9567 20.276V37.6519L40.1711 28.9639L48.9567 20.276ZM48.573 36.7034V21.1486L40.7082 28.9639L48.573 36.7034Z" fill="black" />
+            <path d="M66.1041 55.3446H49.0623L57.5832 46.8415L66.1041 55.3446ZM49.9925 54.9733H65.2483L57.5832 47.3613L49.9925 54.9733Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M92.132 46.8414H91.6674L100.188 55.3445L108.709 46.8414H108.141L107.769 47.2127H107.779L100.188 54.8246L100.165 54.8011L100.137 54.8292L92.132 46.8414Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M74.6256 63.3838V63.8477L83.1465 55.3446L74.6256 46.8415V47.4084L74.9977 47.7797V47.7698L82.6256 55.3446L82.6021 55.3681L82.6301 55.3961L74.6256 63.3838Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M41.07 80.8538L49.3151 72.6019L49.0634 72.3507L40.5425 80.8538H41.07Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M49.0636 55.9237L49.4356 56.294V56.2728L57.0635 63.8476L57.0454 63.8657L57.3062 64.1252L57.5845 63.8476L49.0636 55.3445V55.9237Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M65.8263 72.6283L66.1045 72.3507L57.5836 63.8476V64.4265L57.9557 64.7968V64.7759L65.5836 72.3507L65.5655 72.3688L65.8263 72.6283Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M49.3157 89.1039L49.0623 89.3568L57.5832 97.8599V80.8538L57.2133 81.2229V96.9621L49.3157 89.1039Z" fill="black" />
+            <path fillRule="evenodd" clipRule="evenodd" d="M48.8243 89.1194L40.5413 97.3851V97.8599L49.0623 89.3568L48.8243 89.1194Z" fill="black" />
+            <path d="M40.3708 149.688V138.032H44.2138C46.6078 138.032 48.4348 138.741 48.4348 140.915C48.4348 141.986 47.852 143.072 46.8598 143.435V143.498C48.104 143.797 49.0333 144.695 49.0333 146.254C49.0333 148.601 47.0804 149.688 44.5289 149.688H40.3708ZM42.7019 142.773H44.0564C45.5369 142.773 46.1669 142.19 46.1669 141.245C46.1669 140.222 45.4895 139.828 44.0878 139.828H42.7019V142.773ZM42.7019 147.892H44.3241C45.8991 147.892 46.7654 147.325 46.7654 146.128C46.7654 144.994 45.9306 144.506 44.3241 144.506H42.7019V147.892ZM55.7217 149.908C52.871 149.908 51.1857 148.317 51.1857 144.458V138.032H53.501V144.663C53.501 147.073 54.3515 147.892 55.7217 147.892C57.092 147.892 57.9897 147.073 57.9897 144.663V138.032H60.2262V144.458C60.2262 148.317 58.5725 149.908 55.7217 149.908ZM63.188 149.688V138.032H66.3853C69.9448 138.032 72.134 139.875 72.134 143.813C72.134 147.734 69.9447 149.688 66.5112 149.688H63.188ZM65.5189 147.813H66.2277C68.4169 147.813 69.7557 146.648 69.7557 143.813C69.7557 140.978 68.4169 139.891 66.2277 139.891H65.5189V147.813ZM72.8312 149.688L76.564 138.032H79.3044L83.0372 149.688H80.596L78.9264 143.624C78.5799 142.411 78.2492 141.056 77.9185 139.812H77.8554C77.5404 141.072 77.2096 142.411 76.8631 143.624L75.1937 149.688H72.8312ZM75.1465 146.695V144.899H80.6904V146.695H75.1465ZM84.5747 149.688V138.032H88.4021C90.9692 138.032 92.9537 138.93 92.9537 141.702C92.9537 144.364 90.9692 145.545 88.465 145.545H86.9058V149.688H84.5747ZM86.9058 143.702H88.3076C89.9141 143.702 90.6857 143.025 90.6857 141.702C90.6857 140.332 89.8352 139.891 88.2287 139.891H86.9058V143.702ZM95.2374 149.688V138.032H102.388V139.97H97.5684V142.663H101.648V144.632H97.5684V147.734H102.561V149.688H95.2374ZM108.566 149.908C106.991 149.908 105.447 149.309 104.313 148.223L105.652 146.616C106.471 147.388 107.589 147.892 108.613 147.892C109.826 147.892 110.456 147.372 110.456 146.569C110.456 145.703 109.747 145.419 108.707 144.978L107.148 144.317C105.983 143.844 104.786 142.883 104.786 141.198C104.786 139.292 106.471 137.828 108.818 137.828C110.172 137.828 111.527 138.363 112.472 139.324L111.306 140.789C110.55 140.174 109.778 139.828 108.818 139.828C107.794 139.828 107.148 140.285 107.148 141.041C107.148 141.875 107.983 142.19 108.975 142.584L110.503 143.214C111.905 143.797 112.834 144.71 112.834 146.38C112.834 148.286 111.243 149.908 108.566 149.908ZM116.94 149.688V139.97H113.648V138.032H122.563V139.97H119.271V149.688H116.94Z" fill="black" />
+            <path d="M41.489 166.829C40.3955 166.829 39.6665 166.394 39.1603 165.523L40.193 164.754C40.4765 165.3 40.8714 165.533 41.2966 165.533C41.9649 165.533 42.3091 165.189 42.3091 164.156V159.195H43.8076V164.278C43.8076 165.685 43.0888 166.829 41.489 166.829ZM47.1829 166.829C45.9578 166.829 45.4211 165.989 45.4211 164.561V161.028H46.9095V164.379C46.9095 165.25 47.1525 165.553 47.7094 165.553C48.1751 165.553 48.4789 165.351 48.8636 164.845V161.028H50.3419V166.688H49.1269L49.0155 165.877H48.9851C48.489 166.455 47.9524 166.829 47.1829 166.829ZM51.9388 166.688V161.028H53.1437L53.2652 161.777H53.2955C53.7815 161.291 54.3688 160.886 55.1383 160.886C56.3533 160.886 56.8899 161.736 56.8899 163.154V166.688H55.4015V163.346C55.4015 162.455 55.1484 162.162 54.6017 162.162C54.1359 162.162 53.8423 162.384 53.4272 162.789V166.688H51.9388ZM60.8956 166.829C59.306 166.829 58.0707 165.736 58.0707 163.863C58.0707 162.02 59.3667 160.886 60.7032 160.886C62.2726 160.886 63.0725 162.02 63.0725 163.589C63.0725 163.873 63.0421 164.136 63.0117 164.278H59.5085C59.6401 165.209 60.2476 165.685 61.088 165.685C61.5436 165.685 61.9486 165.543 62.3637 165.29L62.8599 166.212C62.303 166.596 61.5942 166.829 60.8956 166.829ZM59.4882 163.285H61.8069C61.8069 162.516 61.4626 162.03 60.7336 162.03C60.1362 162.03 59.6097 162.455 59.4882 163.285ZM66.2276 166.688V165.837C68.1817 163.964 69.4676 162.698 69.4676 161.605C69.4676 160.866 69.0727 160.42 68.3539 160.42C67.797 160.42 67.3616 160.795 66.9769 161.2L66.1567 160.39C66.8452 159.651 67.5236 159.256 68.5462 159.256C69.9536 159.256 70.9054 160.147 70.9054 161.534C70.9054 162.799 69.7207 164.126 68.3842 165.523C68.7589 165.483 69.255 165.432 69.6094 165.432H71.2496V166.688H66.2276ZM74.9639 166.829C73.5363 166.829 72.2605 165.746 72.2605 163.235C72.2605 160.45 73.6578 159.256 75.1664 159.256C76.0878 159.256 76.756 159.64 77.2015 160.096L76.4118 160.997C76.1485 160.693 75.6929 160.461 75.2575 160.461C74.3564 160.461 73.6274 161.149 73.6274 163.235C73.6274 165.047 74.1944 165.705 74.9234 165.705C75.5005 165.705 75.9562 165.27 75.9562 164.409C75.9562 163.589 75.5309 163.215 74.8829 163.215C74.4779 163.215 74.0122 163.437 73.6274 164.075L73.5667 163.063C73.9717 162.496 74.65 162.162 75.1563 162.162C76.4118 162.162 77.3129 162.88 77.3129 164.409C77.3129 165.898 76.2397 166.829 74.9639 166.829ZM78.2327 164.328V163.265H81.007V164.328H78.2327ZM81.9194 166.688V165.837C83.8735 163.964 85.1594 162.698 85.1594 161.605C85.1594 160.866 84.7645 160.42 84.0456 160.42C83.4888 160.42 83.0534 160.795 82.6686 161.2L81.8485 160.39C82.537 159.651 83.2154 159.256 84.238 159.256C85.6454 159.256 86.5971 160.147 86.5971 161.534C86.5971 162.799 85.4125 164.126 84.076 165.523C84.4506 165.483 84.9468 165.432 85.3011 165.432H86.9414V166.688H81.9194ZM90.4431 166.829C89.0053 166.829 87.9422 166.009 87.9422 164.835C87.9422 163.923 88.5396 163.346 89.2179 162.982V162.941C88.6611 162.526 88.2257 161.989 88.2257 161.23C88.2257 160.015 89.1876 159.256 90.4937 159.256C91.8201 159.256 92.6706 160.045 92.6706 161.24C92.6706 161.949 92.1846 162.536 91.7087 162.87V162.921C92.3871 163.306 92.9642 163.852 92.9642 164.835C92.9642 165.969 91.9517 166.829 90.4431 166.829ZM90.9189 162.516C91.2632 162.141 91.4556 161.757 91.4556 161.331C91.4556 160.724 91.0809 160.299 90.4633 160.299C89.9469 160.299 89.5419 160.623 89.5419 161.23C89.5419 161.898 90.1393 162.202 90.9189 162.516ZM90.4734 165.776C91.1214 165.776 91.5771 165.422 91.5771 164.774C91.5771 164.045 90.8683 163.792 89.9166 163.387C89.5116 163.701 89.2179 164.146 89.2179 164.662C89.2179 165.341 89.7748 165.776 90.4734 165.776ZM94.2789 168.854L93.9751 168.044C94.7649 167.751 95.2002 167.214 95.1901 166.617L95.1395 165.564L95.666 166.434C95.5141 166.586 95.3116 166.657 95.0889 166.657C94.6029 166.657 94.1574 166.333 94.1574 165.776C94.1574 165.24 94.5927 164.895 95.1294 164.895C95.8077 164.895 96.1925 165.442 96.1925 166.353C96.1925 167.548 95.5141 168.449 94.2789 168.854ZM99.4503 166.688V165.837C101.404 163.964 102.69 162.698 102.69 161.605C102.69 160.866 102.295 160.42 101.577 160.42C101.02 160.42 100.584 160.795 100.2 161.2L99.3794 160.39C100.068 159.651 100.746 159.256 101.769 159.256C103.176 159.256 104.128 160.147 104.128 161.534C104.128 162.799 102.943 164.126 101.607 165.523C101.982 165.483 102.478 165.432 102.832 165.432H104.472V166.688H99.4503ZM107.984 166.829C106.435 166.829 105.443 165.493 105.443 163.012C105.443 160.531 106.435 159.256 107.984 159.256C109.513 159.256 110.515 160.542 110.515 163.012C110.515 165.493 109.513 166.829 107.984 166.829ZM107.984 165.665C108.632 165.665 109.108 165.047 109.108 163.012C109.108 160.977 108.632 160.41 107.984 160.41C107.336 160.41 106.85 160.977 106.85 163.012C106.85 165.047 107.336 165.665 107.984 165.665ZM111.395 166.688V165.837C113.349 163.964 114.635 162.698 114.635 161.605C114.635 160.866 114.24 160.42 113.521 160.42C112.964 160.42 112.529 160.795 112.144 161.2L111.324 160.39C112.012 159.651 112.691 159.256 113.713 159.256C115.121 159.256 116.072 160.147 116.072 161.534C116.072 162.799 114.888 164.126 113.551 165.523C113.926 165.483 114.422 165.432 114.776 165.432H116.417V166.688H111.395ZM120.344 166.688V162.283C120.344 161.848 120.384 161.149 120.404 160.714H120.364C120.171 161.109 119.969 161.483 119.756 161.898L118.602 163.731H122.622V164.875H117.154V163.852L119.939 159.387H121.731V166.688H120.344Z" fill="black" />
+          </svg>
+        </Link>
+      </svg>
+    </Link >
+  )
+}
+
+function Footer() {
+  const trianglePaths = [
+    "M160 205.5H159.5V206V240.5V241.707L160.354 240.854L194.854 206.354L195.707 205.5H194.5H160Z",
+    "M161.957 117.644L161.604 117.291L161.25 117.644L136.864 142.03L136.011 142.884L137.218 142.884H185.989L187.196 142.884L186.343 142.03L161.957 117.644Z",
+    "M72.5 244.5L73 244.5L73 244L73 220.5L73 219.274L72.1427 220.15L49.1427 243.65L48.311 244.5L49.5 244.5L72.5 244.5Z",
+    "M118 206L118 206.5L118.5 206.5L142 206.5L143.207 206.5L142.354 205.646L118.854 182.146L118 181.293L118 182.5L118 206Z",
+    "M151.396 185.484L151.75 185.838L152.104 185.484L184.984 152.604L185.838 151.75L184.63 151.75L118.87 151.75L117.662 151.75L118.516 152.604L151.396 185.484Z",
+    "M118 244L118 244.5L118.5 244.5L130 244.5L131.172 244.5L130.361 243.654L118.861 231.654L118 230.756L118 232L118 244Z",
+    "M75 212.5L75 213L75.5 213L87.5 213L88.7444 213L87.846 212.139L75.846 200.639L75 199.828L75 201L75 212.5Z",
+    "M177 115L177 114.5L176.5 114.5L165.5 114.5L164.293 114.5L165.146 115.354L176.146 126.354L177 127.207L177 126L177 115Z",
+    "M179 112L179 112.5L179.5 112.5L202.5 112.5L203.726 112.5L202.85 111.643L179.85 89.1426L179 88.3114L179 89.5L179 112Z",
+    "M198 26.5L198.5 26.5L198.5 26L198.5 3L198.5 1.79289L197.646 2.64645L174.646 25.6464L173.793 26.5L175 26.5L198 26.5Z",
+    "M200.5 181.5V181H200H181.5H180.293L181.146 181.854L199.646 200.354L200.5 201.207V200V181.5Z",
+    "M199.5 244L200 244L200 243.5L200 209L200 207.793L199.146 208.646L164.646 243.146L163.793 244L165 244L199.5 244L199.5 244Z",
+    "M151 66.5L151 67L151.5 67L175 67L176.226 67L175.35 66.1427L151.85 43.1427L151 42.311L151 43.5L151 66.5Z",
+    "M128.5 68H128V68.5V79.5V80.6366L128.838 79.8686L140.838 68.8686L141.785 68H140.5H128.5Z",
+
+  ];
+  return (
+    <svg className={styles.footer_polygons} viewBox="0 0 205 245" fill="none" xmlns="http://www.w3.org/2000/svg">
+      {getTrianglesWithColorAndFill(trianglePaths)}
+    </svg>
+  )
+}
+
+function FooterSeparator() {
+  const trianglePaths = [
+    "M97.6479 100.847L56.702 59.9011L138.594 59.9011L97.6479 100.847Z",
+    "M102.75 23.4079L125.593 46.2513H79.9066L102.75 23.4079Z",
+    "M73.6908 27.3047L57.5755 43.4199L57.5755 11.1895L73.6908 27.3047Z",
+    "M33.0104 34.4541L41.9756 25.4889L41.9756 43.4193L33.0104 34.4541Z",
+    "M12.0981 17.3968L7.72751 13.0262H16.4688L12.0981 17.3968Z",
+    "M44.402 64.7012L48.7726 60.3305V69.0718L44.402 64.7012Z",
+    "M150.171 38.459L143.018 45.6122V31.3058L150.171 38.459Z",
+    "M162.192 26.4369L155.039 19.2836L169.346 19.2836L162.192 26.4369Z",
+    "M174.277 27.3463L192.598 45.6676L155.956 45.6676L174.277 27.3463Z",
+  ]
+  return (
+    <svg className={styles.footer_separator_polygons} viewBox="0 0 194 102" fill="none" xmlns="http://www.w3.org/2000/svg">
+      {getTrianglesWithColorAndFill(trianglePaths)}
+    </svg>
+
+  )
+}
+
+type PolygonsProps = {
+  type: "header-background-desktop" | "header-background-mobile" | "footer" | "footer-separator";
+}
+
+function Polygons({ type: elementType }: PolygonsProps) {
+  let Element: React.ElementType;
+
+  switch (elementType) {
+    case "header-background-desktop":
+      Element = HeaderBackgroundDesktop;
+      break;
+    case "header-background-mobile":
+      Element = HeaderBackgroundMobile;
+      break;
+    case "footer":
+      Element = Footer;
+      break;
+    case "footer-separator":
+      Element = FooterSeparator;
+      break;
+    default:
+      throw new Error("Invalid type provided to Polygons component");
+  }
+
+  return <Element />;
+}
+
+export default Polygons;
